# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema POS",
        "description": "Documentación de la API del Sistema POS",
        "contact": {
            "email": "admin@sistema.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auditoria": {
            "get": {
                "tags": [
                    "Auditoría"
                ],
                "summary": "Obtener registros de auditoría",
                "operationId": "e2a307c9d24225fa07d5af3f86586e31",
                "parameters": [
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "description": "Fecha inicial para filtrar (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "description": "Fecha final para filtrar (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Tipo de auditoría",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "usuario_id",
                        "in": "query",
                        "description": "ID del usuario para filtrar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de registros de auditoría",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "usuario_id": {
                                                        "type": "integer"
                                                    },
                                                    "tipo": {
                                                        "type": "string"
                                                    },
                                                    "tabla": {
                                                        "type": "string"
                                                    },
                                                    "registro_id": {
                                                        "type": "integer"
                                                    },
                                                    "valores_anteriores": {
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "valores_nuevos": {
                                                        "type": "object"
                                                    },
                                                    "ip": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "usuario": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auditoria/apertura-caja": {
            "post": {
                "tags": [
                    "Auditoría"
                ],
                "summary": "Registrar apertura de caja",
                "operationId": "1fc0b617ee03f9528548acac712c8520",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "monto_inicial"
                                ],
                                "properties": {
                                    "monto_inicial": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "observacion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Caja abierta exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "usuario_id": {
                                                    "type": "integer"
                                                },
                                                "fecha_apertura": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "monto_inicial": {
                                                    "type": "number"
                                                },
                                                "monto_final": {
                                                    "type": "number"
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "ABIERTA"
                                                },
                                                "observacion": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación o ya existe una caja abierta"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auditoria/cierre-caja": {
            "post": {
                "tags": [
                    "Auditoría"
                ],
                "summary": "Registrar cierre de caja",
                "operationId": "42e23f68381ef32a2626525fc4cfc700",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "monto_final"
                                ],
                                "properties": {
                                    "monto_final": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "observacion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Caja cerrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "usuario_id": {
                                                    "type": "integer"
                                                },
                                                "fecha_apertura": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "fecha_cierre": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "monto_inicial": {
                                                    "type": "number"
                                                },
                                                "monto_final": {
                                                    "type": "number"
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "CERRADA"
                                                },
                                                "observacion": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación o no hay caja abierta"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}